requirements.txt
Babel==2.10.3
blinker==1.5
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.4
decorator==5.1.1
Flask==2.1.3
Flask-BabelEx==0.9.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
requests==2.28.0
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.39
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.0
WTForms==3.0.1
zipp==3.8.1
